Tighten DateTimeRangeInput API metadata and stories
