+++ src/js/components/TimeInput/propTypes.js
@@ -10,6 +10,7 @@ if (process.env.NODE_ENV !== 'production') {
     format: PropTypes.oneOf(['12', '24']),
     focusPopupOnMount: PropTypes.bool,
     id: PropTypes.string,
+    inline: PropTypes.bool,
     messages: PropTypes.shape({
       activePeriodValue: PropTypes.string,
       activeSection: PropTypes.string,
@@ -27,6 +28,7 @@ if (process.env.NODE_ENV !== 'production') {
     minuteStep: PropTypes.number,
     name: PropTypes.string,
     onChange: PropTypes.func,
+    onPartialChange: PropTypes.func,
     readOnly: PropTypes.bool,
     showSeconds: PropTypes.bool,
     value: PropTypes.string,
